refactor!: moved common structs to connectors
